+++ core/src/main/java/eu/unicore/ucc/actions/ActionBase.java
@@ -117,9 +117,12 @@ public abstract class ActionBase extends Command {
 			error("Problem setting up security.",ioe);
 			endProcessing(ERROR_SECURITY);
 		}
-		
-		initRegistryClient();
-		
+		try {
+			initRegistryClient();
+		}catch(Exception re) {
+			error("Problem setting up registry.", re);
+			endProcessing(ERROR_CONNECTION);
+		}
 		setOutputLocation();
 		
 		String blacklistP = properties.getProperty("blacklist", null);
@@ -144,11 +147,6 @@ public abstract class ActionBase extends Command {
 	public UCCConfigurationProvider getConfigurationProvider() {
 		return configurationProvider;
 	}
-	
-	public void setUCCClientConfigurationProvider(UCCConfigurationProvider configurationProvider) {
-		this.configurationProvider = configurationProvider;
-	}
-	
 
 	/**
 	 * Function that returns true if the connection to registry is not required.
@@ -174,15 +172,12 @@ public abstract class ActionBase extends Command {
 	 * @throws IOException
 	 */
 	public void initConfigurationProvider()throws Exception{
-		if (configurationProvider != null) //unit testing
-			return;
-		String authNMethod = getOption(OPT_AUTHN_METHOD_LONG, OPT_AUTHN_METHOD, UsernameAuthN.NAME);
+		authNMethod = getOption(OPT_AUTHN_METHOD_LONG, OPT_AUTHN_METHOD, UsernameAuthN.NAME);
 		acceptAllIssuers = getBooleanOption(OPT_AUTHN_ACCEPT_ALL_LONG, OPT_AUTHN_ACCEPT_ALL);
 		if(acceptAllIssuers){
 			verbose("Accepting all server CA certificates.");
 		}
 		configurationProvider = new UCCConfigurationProviderImpl(authNMethod, properties, this, acceptAllIssuers);
-		this.authNMethod = authNMethod;
 	}
 
 	/**
@@ -190,7 +185,7 @@ public abstract class ActionBase extends Command {
 	 * commandline or the properties file. If no registry is configured, registry connection
 	 * will not be done. This may or not lead to issues later, depending on the type of command
 	 */
-	protected void initRegistryClient(){
+	protected void initRegistryClient() throws Exception {
 		registryURL=getCommandLine().getOptionValue(OPT_REGISTRY, properties.getProperty(OPT_REGISTRY_LONG));
 		if(registryURL==null || registryURL.trim().length()==0){
 			verbose("No registry is configured.");
@@ -204,48 +199,28 @@ public abstract class ActionBase extends Command {
 			verbose("Registry connection will be skipped.");
 			return;
 		}
-		verbose("Registry = "+registryURL);
 		//accept list of registries either comma- or space-separated
 		String[] urls=registryURL.split("[, ]");
 		if(urls.length>1){
-			MultiRegistryClient erc= new MultiRegistryClient();
+			MultiRegistryClient erc = new MultiRegistryClient();
 			for(String url: urls){
 				if(url.trim().length()==0)continue;
 				verbose("Registry = "+url);
-				try{
-					IRegistryClient c = makeRegistry(url);
-					erc.addRegistry(c);
-				}catch(Exception e){
-					error("Cannot contact registry <"+url+">",null);
-				}	
+				erc.addRegistry(makeRegistry(url));
 			}
 			registry=erc;
 		}
 		else{
-			try{
-				registry = makeRegistry(registryURL);
-				testRegistryConnection();
-			}catch(Exception e){
-				error("Cannot contact registry",e);
-				endProcessing(ERROR_CONNECTION);
-			}
+			verbose("Registry = "+registryURL);
+			registry = makeRegistry(registryURL);
 		}
+		testRegistryConnection();
 	}
 
 	protected IRegistryClient makeRegistry(String url)throws Exception{
-		if(url.contains("/services/Registry")){
-			url = convertToREST(url);
-			verbose("Using converted Registry URL "+url);
-		}
 		IClientConfiguration sec = configurationProvider.getClientConfiguration(url);
 		IAuthCallback auth = configurationProvider.getRESTAuthN();
 		return new RegistryClient(url, sec, auth);
-	}
-
-	private String convertToREST(String soapURL) {
-		String base = soapURL.split("/services/")[0];
-		String regID = soapURL.split("res=")[1]; 
-		return base+"/rest/registries/"+regID;
 	}
 	
 	protected void testRegistryConnection(){