+++ core/src/main/java/eu/unicore/ucc/authn/oidc/OIDCAgentProperties.java
@@ -25,7 +25,8 @@ public class OIDCAgentProperties extends PropertiesHelper {
 	public static final String ACCOUNT = "account";
 	public static final String SCOPE = "scope";
 	public static final String LIFETIME = "lifetime";
-	
+	public static final String REFRESH_INTERVAL = "refreshInterval";
+
 	public static final Map<String, PropertyMD> META = new HashMap<>();
 	static
 	{
@@ -33,6 +34,8 @@ public class OIDCAgentProperties extends PropertiesHelper {
 		META.put(SCOPE, new PropertyMD().setDescription("OpenID scope(s) to request."));
 		META.put(LIFETIME, new PropertyMD().setInt().setPositive().
 				setDescription("Minimum lifetime of the issued access token."));
+		META.put(REFRESH_INTERVAL, new PropertyMD("300").setInt().
+				setDescription("Interval (seconds) before refreshing the token."));
 	}
 
 	public OIDCAgentProperties(Properties properties) throws ConfigurationException